Move advanced IServiceCollection extensions
Moved advanced IServiceCollection extensions methods used primarily by frameworks into a separate namespace: `Microsoft.Framework.DependencyInjection.Extensions`.

Moved:
- `Add(IServiceCollection, ServiceDescriptor)`
- `Add(IServiceCollection, IEnumerable<ServiceDescriptor>)`
- `TryAdd(IServiceCollection, ServiceDescriptor)`
- `TryAdd(IServiceCollection, IEnumerable<ServiceDescriptor>)`
- `TryAdd*`

The default `ServiceCollectionExtensions` now only consists of `AddTransient`, `AddScoped`, `AddSingleton` and `AddInstance` methods.
