+++ test/Microsoft.Framework.DependencyInjection.Tests/ServiceCollectionTests.cs
@@ -1,6 +1,7 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using Microsoft.Framework.DependencyInjection.Extensions;
 using Microsoft.Framework.DependencyInjection.Tests.Fakes;
 using Xunit;
